+++ src/IResultsFormatter.php
@@ -11,11 +11,6 @@ namespace MyTester;
 interface IResultsFormatter
 {
     /**
-     * Anything that needs to be done before tests start running
-     */
-    public function setup(): void;
-
-    /**
      * Report that we started running tests
      *
      * @param TestCase[] $testCases