+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 7.0.0-dev
 - BC break: added method IResultsFormatter::setOutputFileName
 - BC break: added methods ITesterExtension::getEventsBeforeTestCase and getEventsAfterTestCase
 - added TestCaseStarted and TestCaseFinished to public api
+- BC break: removed method IResultsFormatter::setup
 
 Version 6.1.0
 - added CodeCoverageExtension, Reader, IAnnotationsReaderEngine, PhpAttributesEngine, TAssertions::getCounter(), TestsStartedEvent, TestsFinishedEvent to public api